requirements.txt
asgiref==3.4.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Deprecated==1.2.13
Django==3.2.13
django-filter==21.1
django-oauth-toolkit==2.0.0
djangorestframework==3.13.1
drf-writable-nested==0.6.3
idna==3.3
jwcrypto==1.3.1
oauthlib==3.2.0
Pillow==8.4.0
pycparser==2.21
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.1
